Fixed up some warnings found in Windows.

+++ core/XdmfCoreReader.cpp
@@ -165,7 +165,7 @@ public:
           const size_t contentSize = std::strlen(content);
           bool whitespace = true;
 
-          for(int i=0; i<contentSize; ++i) {
+          for(size_t i=0; i<contentSize; ++i) {
             if(!isspace(content[i])) {
               whitespace = false;
               break;
